Refactor and other stuff
