+++ packages/components/src/image_zoom/constants.jsx
@@ -1,2 +1,3 @@
 export const ASPECT_RATIO = 16 / 9;
 export const MAX_SCALE_FACTOR = 4;
+export const CONTAINER_WIDTH_CHANGE_RATE = 500;
