+++ packages/components/src/image_zoom/hooks.jsx
@@ -1,5 +1,14 @@
-import { useEffect, useLayoutEffect, useMemo, useRef, useState } from 'react';
-import { getDistanceBetweenPoints, getElementWidth, getMidpoint, getOffsetForMovement } from './utils';
+import { useCallback, useEffect, useMemo, useReducer, useRef, useState } from 'react';
+import useDebouncedCallback from 'nebenan-react-hocs/lib/use_debounced_callback';
+import {
+  between, getDefaultScale,
+  getDistanceBetweenPoints,
+  getElementWidth,
+  getInsideBoundaries,
+  getMidpoint,
+  getOffsetForMovement, getOffsetForNewScaleWithCustomAnchor
+} from './utils';
+import { CONTAINER_WIDTH_CHANGE_RATE, MAX_SCALE_FACTOR } from './constants';
 
 export const usePrevious = (value) => {
   const ref = useRef(null);
@@ -12,7 +21,7 @@ export const useImage = (src) => {
   const [image, setImage] = useState(null);
 
   useEffect(() => {
-    const imageObj = new Image(100, 100);
+    const imageObj = new Image(100, 100); // width and height get ignored
     imageObj.src = src;
     imageObj.onload = () => { setImage(imageObj); };
   }, [src]);
@@ -20,28 +29,26 @@ export const useImage = (src) => {
   return image;
 };
 
-const useClientLayoutEffect = typeof window !== 'undefined' ? useLayoutEffect : () => {};
-
-export const useContainerWidth = (ref) => {
+const useContainerWidth = (ref) => {
   const [width, setWidth] = useState(getElementWidth(ref?.current));
 
-  const handleResize = () => {
+  const handleResize = useDebouncedCallback(() => {
     if (!ref.current) return;
 
     setWidth(getElementWidth(ref.current));
-  };
+  }, CONTAINER_WIDTH_CHANGE_RATE);
 
-  useClientLayoutEffect(() => {
+  useEffect(() => {
     const el = ref.current;
     if (!el) return;
 
     handleResize();
 
-    const resizeObserver = new ResizeObserver(() => handleResize);
+    const resizeObserver = new ResizeObserver(() => handleResize());
     resizeObserver.observe(el);
 
     return () => { resizeObserver.disconnect(); };
-  }, [ref.current]);
+  }, [ref.current, handleResize]);
 
   return width;
 };
@@ -113,4 +120,114 @@ export const useDrag = (onUpdate) => {
   };
 
   return { start, move, stop };
+};
+
+const TYPE_RESET = 'reset';
+const TYPE_ANCHOR_ZOOM = 'anchor-zoom';
+const TYPE_SAFE_SET_OFFSET = 'safe-set-offset';
+
+export const useImageView = ({ previewSize, imageSize }) => {
+  const getDefaultState = useCallback(() => {
+    let scale = 1;
+    let offset = { top: 0, left: 0 };
+
+    if (previewSize && imageSize) {
+      scale = getDefaultScale(previewSize, imageSize);
+
+      offset = {
+        left: getInsideBoundaries(previewSize.width, imageSize.width * scale, 0),
+        top: getInsideBoundaries(previewSize.height, imageSize.height * scale, 0),
+      };
+    }
+
+    return {
+      scale,
+      offset,
+      defaultScale: scale,
+      maxScale: scale * MAX_SCALE_FACTOR,
+      pristine: true,
+    };
+  }, [previewSize, imageSize]);
+
+  const reducer = useCallback((state, action) => {
+    switch (action.type) {
+      case TYPE_RESET:
+        return getDefaultState();
+
+      case TYPE_ANCHOR_ZOOM: {
+        const { zoomFactor, anchor } = action.payload;
+
+        const scale = between(state.defaultScale, state.maxScale, state.scale * zoomFactor);
+
+        const left = getOffsetForNewScaleWithCustomAnchor(
+          anchor.x,
+          state.offset.left,
+          state.scale,
+          scale,
+          previewSize.width,
+        );
+
+        const top = getOffsetForNewScaleWithCustomAnchor(
+          anchor.y,
+          state.offset.top,
+          state.scale,
+          scale,
+          previewSize.height,
+        );
+
+        const offset = { top, left };
+
+        return { ...state, scale, offset, pristine: false };
+      }
+
+      case TYPE_SAFE_SET_OFFSET: {
+        const { offset } = action.payload;
+
+        const left = getInsideBoundaries(
+          previewSize.width,
+          imageSize.width * state.scale,
+          offset.left,
+        );
+
+        const top = getInsideBoundaries(
+          previewSize.height,
+          imageSize.height * state.scale,
+          offset.top,
+        );
+
+        return { ...state, offset: { left, top }, pristine: false };
+      }
+      default: return state;
+    }
+  }, [previewSize, imageSize, getDefaultState]);
+
+  const [state, dispatch] = useReducer(reducer, getDefaultState());
+
+  const actions = {
+    reset: () => dispatch(({
+      type: TYPE_RESET,
+    })),
+    safeSetOffset: (offset) => dispatch(({
+      type: TYPE_SAFE_SET_OFFSET,
+      payload: { offset },
+    })),
+    anchorZoom: (zoomFactor, anchor) => dispatch(({
+      type: TYPE_ANCHOR_ZOOM,
+      payload: { zoomFactor, anchor },
+    })),
+  };
+
+  const previousImageSize = usePrevious(imageSize);
+  const previousPreviewSize = usePrevious(previewSize);
+  useEffect(() => {
+    if (!imageSize || !previewSize) return;
+
+    const isChangedImage = imageSize !== previousImageSize;
+    const isChangedPristinePreview = previewSize !== previousPreviewSize && state.pristine;
+    if (!isChangedImage && !isChangedPristinePreview) return;
+
+    actions.reset();
+  }, [previewSize, imageSize]);
+
+  return [state, actions];
 };
